refactoring: features moved to Home page component
