+++ src/page/Home/Slider/Card/Card.css
