+++ src/page/Home/Slider/Card/Card.sass
