+++ src/page/Home/Slider/Slider.sass
