+++ src/page/Home/Tab/SliderTab/CodeExamples/index.ts
