+++ src/page/Home/Tab/TabTab/TabTab.tsx
@@ -1,4 +1,4 @@
-import CodeBlock from "../../../shared/CodeBlock/CodeBlock";
+import CodeBlock from "../../../../shared/CodeBlock/CodeBlock";
 import {
 	OpenCloseTab,
 	states,
