+++ src/page/Home/Tab/ThemeTab/ThemeTab.tsx
@@ -1,4 +1,4 @@
-import CodeBlock from "../../../shared/CodeBlock/CodeBlock";
+import CodeBlock from "../../../../shared/CodeBlock/CodeBlock";
 import { states, effects, markup, imports } from "./CodeExamples";
 
 const ThemeTab = () => {
